Update AutoLineBudget.cs
Line 192. Equation for newOccupancy was incorrect. Why would it depend on the line.m_budget value? The regulation worked good only in the beginning (when line.m_budget == 100). Later it got stuck around some value.
Line 197. Equation for vehChange was incorrect. As a result inertia worked not as supposed.
Conclusion: Every code must be tested!!!